[REFACTOR] Move delete instruction article flow into new component
